+++ src/canceler.ts
@@ -1,4 +1,4 @@
-import { Listener } from './agent'
+import { Handler } from './agent'
 import { IpcService } from './aliases'
 
 /**
@@ -23,14 +23,14 @@ export class Canceler implements Cancelable {
    * Initializes a new Canceler.
    * @param ipcService The IPC service to use for unsubscribing
    * @param channel The channel to unsubscribe from
-   * @param listener The listener to unsubscribe
+   * @param handler The handler to unsubscribe
    */
-  constructor (private ipcService: IpcService, public channel: string, private listener: Listener) {}
+  constructor (private ipcService: IpcService, public channel: string, private handler: Handler) {}
 
   /**
    * Unsubscribes the listener, meaning it will no longer be called when a message is received.
    */
   public cancel (): void {
-    this.ipcService.removeListener(this.channel, this.listener)
+    this.ipcService.removeListener(this.channel, this.handler)
   }
 }