+++ src/client.ts
@@ -1,5 +1,5 @@
 import { Agent } from './agent'
-import { IpcRenderer, IpcRendererEvent } from './aliases'
+import { IpcRenderer } from './aliases'
 
 /**
  * Represents an API wrapper around Electron's ipcRenderer.
@@ -14,10 +14,10 @@ export class Client extends Agent<IpcRenderer> {
 
   /**
    * Sends a message to the main process.
-   * @param channel The channel to use for sending the request
-   * @param data The request data
+   * @param channel The channel to use for sending the data
+   * @param data The data to send
    */
-  protected send (channel: string, ...data: any[]): void {
-    this.ipcService.send(channel, ...data)
+  protected send (channel: string, data: any): void {
+    this.ipcService.send(channel, data)
   }
 }